add entities utils module
add module for housing entities utility functions
